backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.11.3

SQLAlchemy==2.0.36
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic[email]